requirements.txt
pytest~=7.4.0
setuptools~=69.1.1
SQLAlchemy~=2.0.15
PyYAML~=6.0.1